Hauntify/requirements.txt
streamlit>=1.25.0
opencv-python>=4.8.0
numpy>=1.25.0
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
